requirements.txt
# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LiveKit
livekit-api>=0.8.0
livekit-agents[xai]>=1.3.0